extracted dashboard form into a component to avoid error on npm run build
